chore: split TraceViewer into files, small refactorings (#178)
